fix(phase0): review-driven hardening for Phase 1 handoff
Three fixes from the branch-level code review:

- V2Policy.new_state: delegate to each layer's new_state() instead of
  asserting DenseLayer. The existing assertion would fire the moment
  Phase 1 adds GruLayer and masquerade as an unrelated regression.
  forward() and new_state() signatures relax from list[None] to list[Any]
  since Phase 1+ layer states are not None.

- model_io.load_policy_from_json: load weights at f64 and let torch.copy_
  cast to the destination dtype. Previously hardcoded f32, silently
  downcasting f64 JSON weights (the exporter writes f64 when the source
  policy is .double()'d, as in the cross-language equivalence test).

- v2 parser row-length mismatch error: include row index + expected vs
  actual column counts. Previously "Layer N weight row length mismatch"
  gave no actionable info.

+++ src/python/aerocapture/training/model_io.py
@@ -34,8 +34,11 @@ def load_policy_from_json(path: str, device: str | torch.device) -> V2Policy:
         if layer_spec.type == "dense":
             if lw.w is None or lw.b is None:
                 raise ValueError(f"Dense layer {key} missing w/b in {path}")
-            w = torch.tensor(lw.w, dtype=torch.float32, device=device)
-            b = torch.tensor(lw.b, dtype=torch.float32, device=device)
+            # JSON stores Python floats (f64). Load at f64 and let `.copy_`
+            # cast to the destination policy's dtype -- preserves precision
+            # for f64 policies, safely downcasts for f32 policies.
+            w = torch.tensor(lw.w, dtype=torch.float64, device=device)
+            b = torch.tensor(lw.b, dtype=torch.float64, device=device)
             layer = policy.layers[i]
             assert isinstance(layer, DenseLayer)
             with torch.no_grad():
